Submenu now shows up. Also refactored Submenu, SubMenuDropdown.
