+++ src/ListItem.js
@@ -1,7 +1,16 @@
 import "./ListItem.css";
 import SubMenu from "./SubMenu";
+import Backdrop from "./Backdrop";
+import SubMenuToggle from "./SubMenuToggle";
+import { useState } from "react";
 
 function ListItem(props) {
+  const [dropDown, setDropDown] = useState(false);
+
+  function handleDropDown() {
+    setDropDown(!dropDown);
+  }
+
   return (
     <li className={props.checked ? "done" : ""}>
       <input
@@ -13,12 +22,19 @@ function ListItem(props) {
       />
       <label htmlFor={props.id}>{props.text}</label>
       <span className="dot">{priorityToIcon[props.priority]}</span>
-      <SubMenu
-        priority={props.priority}
-        onChangePriority={props.onChangePriority}
-        onDeleteTask={props.onDeleteTask}
-        id={props.id}
-      />
+      <SubMenuToggle onToggle={handleDropDown} />
+      {dropDown ? (
+        <>
+          <Backdrop onClickBackdrop={handleDropDown} />
+          <SubMenu
+            priority={props.priority}
+            onChangePriority={props.onChangePriority}
+            onDeleteTask={props.onDeleteTask}
+            onHandleDropDown={handleDropDown}
+            id={props.id}
+          />
+        </>
+      ) : null}
     </li>
   );
 }
