+++ src/SubMenuToggle.js
@@ -1,0 +1,15 @@
+function SubMenuToggle(props) {
+    return <svg
+    xmlns="http://www.w3.org/2000/svg"
+    className="bi bi-list dropbtn"
+    viewBox="0 0 16 16"
+    onClick={props.onToggle}
+  >
+    <path
+      fillRule="evenodd"
+      d="M2.5 12a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5zm0-4a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5zm0-4a.5.5 0 0 1 .5-.5h10a.5.5 0 0 1 0 1H3a.5.5 0 0 1-.5-.5z"
+    />
+  </svg>
+}
+
+export default SubMenuToggle;
